Move helpers out of editor init file
